Refactor some more files
